Object 재활용
게임 루프 안에서 오브젝트를 생성, 삭제하는 것은 좋지 않으므로 재활용을 함
만약 재활용 할 수 없다면 생성하고, 할 수 있다면 가져다 쓰는 방식으로 처리

재활용 하는 객체의 판단은 Recyclable 인터페이스로 판단
Class 를 Key로 하는 Hashmap 으로 재활용 클래스를 관리함

+++ 6. DragonFlight/app/src/main/java/kr/ac/tukorea/sgp/s2018182024/dragonflight/framework/AnimationSprite.java
@@ -8,7 +8,7 @@ public class AnimationSprite extends Sprite {
     protected int frameCount;
     protected int frameWidth, frameHeight;
     protected float fps;
-    protected float createdTime;
+    protected long createdTime;
 
     public AnimationSprite(int resId, float x, float y, float width, float height, float fps, int frameCount) {
         super(resId, x, y, width, height);
